+++ apps/doc/src/DocViewer.tsx
@@ -28,9 +28,10 @@ import {DocFindButton} from "./DocFindButton";
 import IconButton from "@material-ui/core/IconButton";
 import MenuIcon from '@material-ui/icons/Menu';
 import {MUIPaperToolbar} from "../../../web/js/mui/MUIPaperToolbar";
-import {DocRenderer} from "./renderers/DocRenderer";
+import {DocRenderer, DocViewerFileTypeContext} from "./renderers/DocRenderer";
 import {useLogger} from "../../../web/js/mui/MUILogger";
 import { ViewerContainerProvider } from "./ViewerContainerStore";
+import {FileTypes} from "../../../web/js/apps/main/file_loaders/FileTypes";
 
 const Main = React.memo(() => {
 
@@ -293,12 +294,16 @@ export const DocViewer = React.memo(() => {
         return <LoadingProgress/>
     }
 
+    const fileType = FileTypes.create(docURL);
+
     return (
-        <ViewerContainerProvider>
-            <DeviceRouter handheld={<Device.Handheld/>}
-                          desktop={<Device.Desktop/>}/>
-        </ViewerContainerProvider>
-    )
+        <DocViewerFileTypeContext.Provider value={fileType}>
+            <ViewerContainerProvider>
+                <DeviceRouter handheld={<Device.Handheld/>}
+                              desktop={<Device.Desktop/>}/>
+            </ViewerContainerProvider>
+        </DocViewerFileTypeContext.Provider>
+    );
 
 }, isEqual);